+++ app/src/main/java/acr/appcradle/shoppinglist/ui/screens/auth/AuthViewModel.kt
@@ -1,7 +1,7 @@
 package acr.appcradle.shoppinglist.ui.screens.auth
 
-import acr.appcradle.shoppinglist.data.model.AuthResponse
 import acr.appcradle.shoppinglist.data.repository.AuthRepository
+import acr.appcradle.shoppinglist.model.AuthUiState
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -35,11 +35,3 @@ class AuthViewModel @Inject constructor(
     }
 }
 
-sealed class AuthUiState {
-    object Initial : AuthUiState()
-    object Loading : AuthUiState()
-    data class Success(val response: AuthResponse) :
-        AuthUiState()
-
-    data class Error(val message: String) : AuthUiState()
-}
