+++ app/src/main/java/acr/appcradle/shoppinglist/ui/screens/register/RegisterViewModel.kt
@@ -1,7 +1,7 @@
 package acr.appcradle.shoppinglist.ui.screens.register
 
-import acr.appcradle.shoppinglist.data.model.AuthResponse
 import acr.appcradle.shoppinglist.data.repository.AuthRepository
+import acr.appcradle.shoppinglist.model.RegisterUiState
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -35,11 +35,3 @@ class RegisterViewModel @Inject constructor(
     }
 }
 
-sealed class RegisterUiState {
-    object Initial : RegisterUiState()
-    object Loading : RegisterUiState()
-    data class Success(val response: AuthResponse) :
-        RegisterUiState()
-
-    data class Error(val message: String) : RegisterUiState()
-}
